Renaming SAP_Engine/Compute files.
