+++ SAP_Engine/Compute/PushToXML.cs
@@ -19,48 +19,44 @@
  * You should have received a copy of the GNU Lesser General Public License     
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
+
 using System;
 using System.Collections.Generic;
-using System.Text;
-using System.Xml.Serialization;
 using System.IO;
-using BH.oM.Environment.SAP.XML;
-using BH.Engine.Environment.SAP;
+using System.Text;
+using System.Linq;
+using BH.oM.Environment.SAP.Stroma10;
 using BH.oM.Base;
-using BH.oM.Adapter;
+using System.Security.Cryptography.X509Certificates;
+using System.Runtime.InteropServices.ComTypes;
+using System.Xml.Serialization;
 using System.ComponentModel;
 using BH.oM.Base.Attributes;
+using BH.oM.Adapter;
 
 namespace BH.Engine.Environment.SAP
 {
     public static partial class Compute
     {
-        [Description("A test??.")]
-        [Input("fileSettingsInput", "Location of the XML files to pull from.")]
-        [Input("fileSettingsOutput", "Location of the XML files to push to.")]
-        [Input("sAP10Data", "data to push to files.")]
-        [Input("run", "Run the method?.")]
-        [Output("success", "Has the method run?.")]
-        public static bool PushToXMLFile(FileSettings fileSettingsInput, FileSettings fileSettingsOutput, BH.oM.Environment.SAP.XML.SAP10Data sAP10Data , bool run = false)
+        [Description("Converting a SAP object to an XML files.")] 
+        [Input("data", "A SAPReport object to convert to an xml file.")]
+        [Input("fileSettings", "Location of the XML files to push from.")]
+        [Input("run", "Run the method.")]
+        [Output("success","Has the method run.")]
+        public static bool PushToXML(BH.oM.Environment.SAP.XML.SAPReport data, FileSettings fileSettings, bool run = false)
         {
+
             if (!run)
                 return false;
 
             XmlSerializerNamespaces xns = new XmlSerializerNamespaces();
-            XmlSerializer szer = new XmlSerializer(typeof(SAPReport));
+            XmlSerializer szer = new XmlSerializer(typeof(BH.oM.Environment.SAP.XML.SAPReport));
 
-            TextReader tr = new StreamReader(Path.Combine(fileSettingsInput.Directory, fileSettingsInput.FileName));
-            var data = (SAPReport)szer.Deserialize(tr);
-            tr.Close();
-
-            data.SAP10Data = sAP10Data;
-
-            TextWriter tw = new StreamWriter(Path.Combine(fileSettingsOutput.Directory, fileSettingsOutput.FileName));
-            szer.Serialize(tw, data, xns);
-            tw.Close();
+            TextWriter ms = new StreamWriter(Path.Combine(fileSettings.Directory, fileSettings.FileName));
+            szer.Serialize(ms, data, xns);
+            ms.Close();
 
             return true;
         }
     }
 }
-
